requirements.txt
pdfplumber
tiktoken
python-dotenv
chromadb
google-genai
google-adk
